+++ src/three/gltf/classes/StructuralMetadata.js
